Example code cleanup
git-svn-id: http://encog-java.googlecode.com/svn/trunk/encog-examples@1236 f90f6e9a-ac51-0410-b353-d1b83c6f6923

+++ src/org/encog/examples/neural/bam/BidirectionalAssociativeMemory.java
@@ -1,6 +1,30 @@
+/*
+ * Encog Artificial Intelligence Framework v2.x
+ * Java Examples
+ * http://www.heatonresearch.com/encog/
+ * http://code.google.com/p/encog-java/
+ * 
+ * Copyright 2008-2010, Heaton Research Inc., and individual contributors.
+ * See the copyright.txt in the distribution for a full listing of 
+ * individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
 package org.encog.examples.neural.bam;
 
-import org.encog.neural.data.NeuralData;
 import org.encog.neural.data.bipolar.BiPolarNeuralData;
 import org.encog.neural.networks.BasicNetwork;
 import org.encog.neural.networks.NeuralDataMapping;
@@ -43,7 +67,7 @@ public class BidirectionalAssociativeMemory {
 		for(int i=0;i<str.length();i++)
 		{
 			char ch = Character.toUpperCase(str.charAt(i));
-			int idx = (int)ch-FIRST_CHAR;
+			int idx = ch-FIRST_CHAR;
 			
 			int place = 1;
 			for( int j=0;j<BITS_PER_CHAR;j++)
